TOOLS-2805: Add mod-tidy task to Evergreen Static Analysis variant
