+++ build.go
@@ -12,8 +12,13 @@ import (
 var taskRegistry = task.NewRegistry(task.WithAutoNamespaces(true))
 
 func init() {
+	// Build
+	taskRegistry.Declare("build").Description("build the tools").OptionalArgs("tools").Do(buildscript.BuildTools)
+
+	// Static Analysis
 	taskRegistry.Declare("sa:modtidy").Description("runs go mod tidy").Do(buildscript.SAModTidy)
-	taskRegistry.Declare("build").Description("build the tools").DependsOn("sa:modtidy").OptionalArgs("tools").Do(buildscript.BuildTools)
+
+	// Testing
 	taskRegistry.Declare("test:unit").Description("runs unit tests").OptionalArgs("tools").Do(buildscript.TestUnit)
 	taskRegistry.Declare("test:integration").Description("runs integration tests").OptionalArgs("tools", "ssl", "auth", "kerberos", "topology").Do(buildscript.TestIntegration)
 	taskRegistry.Declare("test:kerberos").Description("runs kerberos tests").Do(buildscript.TestKerberos)
